+++ kernax/wrappers/__init__.py
@@ -3,7 +3,6 @@ from .ARDKernel import ARDKernel
 from .BatchKernel import BatchKernel
 from .BlockDiagKernel import BlockDiagKernel
 from .BlockKernel import BlockKernel
-from .DiagKernel import DiagKernel, StaticDiagKernel
 from .ExpKernel import ExpKernel
 from .LogKernel import LogKernel
 from .NegKernel import NegKernel
@@ -11,8 +10,6 @@ from .WrapperKernel import WrapperKernel
 
 __all__ = [
 	"WrapperKernel",
-	"DiagKernel",
-	"StaticDiagKernel",
 	"ExpKernel",
 	"LogKernel",
 	"NegKernel",